Fix build break.
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@50552 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lsass/server/auth-providers/ad-open-provider/adstruct.h
@@ -182,7 +182,7 @@ typedef struct __AD_PROVIDER_CONTEXT
     gid_t gid;
     pid_t pid;
 
-    DWORD nRefCount;
+    LONG nRefCount;
 
     PLSA_AD_PROVIDER_STATE pState;
 } AD_PROVIDER_CONTEXT, *PAD_PROVIDER_CONTEXT;
